Multiple independent Moongen support
When try to run multiple Moongen instance, it failed because it does not
meet the requirement at
http://dpdk.org/doc/guides/prog_guide/multi_proc_support.html#running-multiple-independent-dpdk-applications.

To achieve the multiple MoonGen instance, we need:
a) Different file prefix for the MoonGen instances.
b) Instead of reserving all the available memory, specify the memory
allocation explicitly.
c) Black list the un-used PCI port, so that the second instance will not
reset the PCI device that has been assigned to the first instance.

Signed-off-by: Yunhong Jiang <yunhong.jiang@linux.intel.com>